AspectTypeError: Move and rename class
Import aliasing at test module is required to maintain compatibility
with https://github.com/coala/coala/pull/3894
